Данные(константы) выделены в отдельные модуль
